[OBMIN-31] Created new endpoint for getting all created advertisements by logged in user, extracted UserService's methods into an interface
