+++ src/main/java/space/obminyashka/items_exchange/dao/AdvertisementRepository.java
@@ -11,6 +11,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Collection;
 import java.util.Optional;
 
 @Repository
@@ -38,4 +39,6 @@ public interface AdvertisementRepository extends JpaRepository<Advertisement, Lo
                                                 @Param("subcategoryId") Long subcategoryId,
                                                 @Param("categoryId") Long categoryId,
                                                 @Param("locationId") Long locationId);
+
+    Collection<Advertisement> findAllByUserUsername(String username);
 }
