+++ src/main/java/space/obminyashka/items_exchange/dto/UserDto.java
@@ -13,8 +13,6 @@ import java.util.List;
 import java.util.Set;
 
 @NoArgsConstructor
-@AllArgsConstructor
-@Builder
 @Getter
 @Setter
 public class UserDto {
